added db_print command and some comment fixes
db_print command prints the current tables in the database into files so that we can use them later to load the same database from flat files to sqlserver, or db2, or shore (if someone writes this functionality one day).
to be able to do such loading in parallel in shore, db_print command takes one argument which specifies the maximum number of tuples that can be written to a file and writes a table into multiple files if needed.
if this argument is given as 0, then the whole table is written to a single file.

right now, only tpce uses the functionality invoked by db_print command but it's easy to do this in other benchmarks too as needed. just look at the db_print function in shore_tpce_env.cpp.