tpce xct checks
(1) code cleanup
(2) in trade-lookup, there was a problem with if-else statements; (is-cash if statement should have finished earlier)
(3) in broker_volume, the first part of the harness control doesn't make sense for the initial database
(4) trade_result and market_feed; require trade_order to be executed first and create input for them. the way djordje handled this was to remove these two transactions from the mix and try to have a heuristic of when to execute them but it still caused crashes due to empty input for these two transactions. i simply kept the mix but i abort if the input is null (added another error to shore_error.h for this). ideally, during the miss the abort rate due to this should be very low. i also had to put a lock (in MEESUT.h) for the buffers that keep already created inputs for these two transactions since multiple threads might be accessing to them.